+++ home-assistant-plugin/custom_components/xknx/const.py
@@ -26,7 +26,7 @@ CONF_SYNC_STATE = "sync_state"
 CONF_RESET_AFTER = "reset_after"
 
 ATTR_COUNTER = "counter"
-ATTR_LAST_TELEGRAM_IA = "last_telegram_IA"
+ATTR_SOURCE = "source"
 ATTR_LAST_KNX_UPDATE = "last_knx_update"
 ATTR_MOVEMENT_LOCKED = "movement_locked"