+++ xknx/devices/sensor.py
@@ -77,6 +77,11 @@ class Sensor(Device):
         """Return unique id for this device."""
         return f"{self.sensor_value.group_address_state}"
 
+    @property
+    def last_telegram(self) -> Telegram | None:
+        """Return the last telegram received from the RemoteValue."""
+        return self.sensor_value.telegram
+
     async def process_group_write(self, telegram: "Telegram") -> None:
         """Process incoming and outgoing GROUP WRITE telegram."""
         await self.sensor_value.process(telegram, always_callback=self.always_callback)