+++ changelog.md
@@ -2,20 +2,24 @@
 
 ## Unreleased changes
 
+## Devices
+
 - Add support for cover lock
+- ExposeSensor values can now be read from other xknx devices that share a group address
 - Add more information to sensors and binary sensors in the HA integration
-- Store last Telegram and decoded value in RemoteValue
+
 
 ### Breaking Changes
 
 - Remove configuration handling from core library (use https://xknx.io/config-converter)
-- Drop support for python 3.7
 
 ### Internals
 
+- Drop support for python 3.7
 - use pytest tests instead of unittest TestCase
 - Move RequestResponse and subclasses to xknx.io.request_response.*
 - Move ConnectionConfig to xknx.io.connection
+- Store last Telegram and decoded value in RemoteValue
 
 ## 0.17.5 Add support for unique ids 2021-03-30